requirements.txt
speechrecognition
pyttsx3
python-dotenv
openai
bardapi
pyaudio
requests